Update version in Cargo.toml after feature addition
The version of the rgen package in the Cargo.toml file has been upgraded from 1.1.0 to 2.0.0. This change is made to reflect the addition of a new feature, allowing the password generator to store the output in a file. The 'save' flag now accepts filename as an argument, providing users with an option for automated password storage.