Bump rgen version to 2.0.1 in Cargo.toml
The version of rgen has been bumped from 2.0.0 to 2.0.1 in Cargo.toml. This is to align with the latest code changes which included both refactoring work and improvements to code readability. More specifically, the password generating functionality was moved to a separate module named password_generator.rs, thereby simplifying our main.rs file and allowing it to primarily focus on argument parsing and the calling of the password generation functions.